+++ src/components/secciones/Baterias.js
@@ -1,5 +1,6 @@
 import React, { Component } from "react";
 import history from "../../history";
+import {Helmet} from "react-helmet";
 
 //CSS
 import "../../assets/css/Baterias.css";
@@ -8,7 +9,6 @@ import "../../assets/css/Baterias.css";
 import Header from "../Header";
 import Footer from "../../components/Footer";
 import Scrollchor from "react-scrollchor";
-import SEO from "../elementos/SEO";
 
 //Images
 import flechaAbajo from "../../assets/img/angle-arrow-down.png";
@@ -43,11 +43,14 @@ class Baterias extends Component {
     }
 
     render() {
-		let title = "Venta de baterias";
+		let title = "Venta de baterias - Electrobat";
 		let descripcion = "Venta de baterias para autos Nafteros, GNC, Diesel, también camiones, y tractores de gran porte. Modelos: Caterpillar, Sprinter, Ford K, Ranger, NS 40, N70Z, N75Z, entre otros."
         return (
             <div className="baterias">
-                <SEO title={title} description={descripcion}/>
+                <Helmet>
+					<title>{title}</title>
+					<meta name="description" content={descripcion}></meta>
+				</Helmet>
                 <div className="imagenBaterias imagenFondo">
                     <Header />
                     <div className="item"></div>