+++ src/components/secciones/Volquetes.js
@@ -2,6 +2,7 @@ import React, { Component } from "react";
 import { connect } from "react-redux";
 import history from "../../history";
 import Scrollchor from "react-scrollchor";
+import {Helmet} from "react-helmet";
 
 //CSS
 import "../../assets/css/Volquetes.css";
@@ -9,7 +10,6 @@ import "../../assets/css/Volquetes.css";
 //Components
 import Header from "../Header";
 import Footer from "../../components/Footer";
-import SEO from "../elementos/SEO";
 
 
 //Images
@@ -36,11 +36,14 @@ class Volquetes extends Component {
     }
 
     render() {
-		let title		= "Alquiler de volquetes";
+		let title		= "Alquiler de volquetes - Electrobat";
 		let descripcion = title + ". Rosario y alrededores. Tamaño mediano y grande.";
         return (
             <div className="volquetes">
-                <SEO title={title} description={descripcion} />
+                <Helmet>
+					<title>{title}</title>
+					<meta name="description" content={descripcion}></meta>
+				</Helmet>
                 <div className="imagenVolquetes imagenFondo">
                     <Header />
                     <div className="item"></div>