+++ CHANGELOG.md
@@ -9,7 +9,7 @@ y este proyecto adhiere a [Semantic Versioning]( https://semver.org/spec/v2.0.0.
 -
 
 ### Changed
--
+- SEO: uso de react helmet.
 
 ### Deprecated
 -